Determinization of document generation
Instead of using HashSet (and HahsMap) instances, the JSON schema and
generator implementations use LinkedHashSet and TreeSet to guarantee
that the order will be the same each execution. This means that if the
user provides a Random instance with a fixed seed, the output of the
generator will always be the same. This is useful for debugging and
benchmarking purpose.

Moreover, JSONObject and JSONArray got overriden to also use TreeSet and
TreeMap. Again, the point is to determinize the behavior.