Validator: better handling of const and anyOf
anyOf: instead of applying the same procedure than for the generator,
the validator retrieves the information as-is. That is, the validator
works on the direct contraints stored in the schema, rather than the
transformation constructed for the generator. This is to avoid infinite
loops when constraints can not be simplified and an "anyOf" remains at
the top.

const: possibility to decide whether to check the type or the exact
values. This is useful for const constraints inside a not (due to our
abstractions).